// src/Moryx.Media.Web/src/app/api/models/variant-descriptor.ts
/* tslint:disable */
/* eslint-disable */
import { PreviewDescriptor } from './preview-descriptor';
export interface VariantDescriptor {
  creationDate?: string;
  extension?: null | string;
  fileHash?: null | string;
  mimeType?: null | string;
  name?: null | string;
  preview?: PreviewDescriptor;
  size?: number;
}